moved common code to base class
